requirements.txt
python-dateutil>=2.9,<3.0
pyyaml>=6.0,<7.0
